Updated styling and folder org
