+++ src/components/DogImage/DogImage.css
@@ -1,82 +1,81 @@
 section.dogImageSection {
-    height: 250px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
+  height: 250px;
+  display: flex;
+  grid-column-start: span 2;
 }
 
 div .dogImage {
-    text-align: center;
-    height: 100%;
-    width: auto;
-    margin: auto;
-    padding: 10px;
+  text-align: center;
+  height: 100%;
+  width: auto;
+  margin: auto;
+  padding: 10px;
 }
 
 .dogImage img {
-    height: 90%;
-    width: auto;
-    box-shadow: 3px 3px 13px -2px rgba(0, 0, 0, 0.75);
-    -webkit-box-shadow: 3px 3px 13px -2px rgba(0, 0, 0, 0.75);
-    -moz-box-shadow: 3px 3px 13px -2px rgba(0, 0, 0, 0.75);
-    border-radius: 10px;
+  height: 90%;
+  width: auto;
+  box-shadow: 3px 3px 13px -2px rgba(0, 0, 0, 0.75);
+  -webkit-box-shadow: 3px 3px 13px -2px rgba(0, 0, 0, 0.75);
+  -moz-box-shadow: 3px 3px 13px -2px rgba(0, 0, 0, 0.75);
+  border-radius: 10px;
 }
 .dogImage img:hover {
-    cursor: pointer;
+  cursor: pointer;
 }
 
 .dogImage button {
-    height: 100%;
-    background: none;
-    border: none;
+  height: 100%;
+  background: none;
+  border: none;
 }
 
 .dogImage button {
-    --c: #ede5c3; /* the color*/
+  --c: #ede5c3; /* the color*/
 
-    /* box-shadow: 0 0 0 0.1em inset var(--c); */
-    --_g: linear-gradient(var(--c) 0 0) no-repeat;
-    background: var(--_g) calc(var(--_p, 0%) - 100%) 0%,
+  /* box-shadow: 0 0 0 0.1em inset var(--c); */
+  --_g: linear-gradient(var(--c) 0 0) no-repeat;
+  background: var(--_g) calc(var(--_p, 0%) - 100%) 0%,
     var(--_g) calc(200% - var(--_p, 0%)) 0%,
     var(--_g) calc(var(--_p, 0%) - 100%) 100%,
     var(--_g) calc(200% - var(--_p, 0%)) 100%;
-    background-size: 50.5% calc(var(--_p, 0%) / 2 + 0.5%);
-    outline-offset: 0.1em;
-    transition: background-size 0.4s, background-position 0s 0.4s;
-    border-radius: 10px;
+  background-size: 50.5% calc(var(--_p, 0%) / 2 + 0.5%);
+  outline-offset: 0.1em;
+  transition: background-size 0.4s, background-position 0s 0.4s;
+  border-radius: 10px;
 }
 .dogImage button:hover {
-    --_p: 100%;
-    transition: background-position 0.4s, background-size 0s;
+  --_p: 100%;
+  transition: background-position 0.4s, background-size 0s;
 }
 .dogImage button:active {
-    box-shadow: 0 0 9e9Q inset #0009;
-    background-color: var(--c);
-    color: #fff;
+  box-shadow: 0 0 9e9Q inset #0009;
+  background-color: var(--c);
+  color: #fff;
 }
 
 button.newbtn {
-    margin: auto;
-    border: 3px solid #f56960;
-    background: #daa588;
-    padding: 1em;
-    border-radius: 40px;
-    font-size: 2em;
-    background: rgba(255, 255, 255, 0.2);
-    border-radius: 16px;
-    box-shadow: 0 4px 30px rgb(0 0 0 / 10%);
-    backdrop-filter: blur(5px);
-    -webkit-backdrop-filter: blur(5px);
-    border: 1px solid rgba(255, 255, 255, 0.3);
+  margin: auto;
+  border: 3px solid #f56960;
+  background: #daa588;
+  padding: 1em;
+  border-radius: 40px;
+  font-size: 2em;
+  background: rgba(255, 255, 255, 0.2);
+  border-radius: 16px;
+  box-shadow: 0 4px 30px rgb(0 0 0 / 10%);
+  backdrop-filter: blur(5px);
+  -webkit-backdrop-filter: blur(5px);
+  border: 1px solid rgba(255, 255, 255, 0.3);
 }
 
 button.newbtn:hover {
-    background: rgba(243, 201, 139, 0.5);
-    border-radius: 16px;
-    box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
-    backdrop-filter: blur(14.2px);
-    -webkit-backdrop-filter: blur(14.2px);
-    border: 1px solid rgba(243, 201, 139, 1);
-    cursor: pointer;
-    transition: 0.5s;
+  background: rgba(243, 201, 139, 0.5);
+  border-radius: 16px;
+  box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
+  backdrop-filter: blur(14.2px);
+  -webkit-backdrop-filter: blur(14.2px);
+  border: 1px solid rgba(243, 201, 139, 1);
+  cursor: pointer;
+  transition: 0.5s;
 }
