+++ src/pages/CatFactPage/CatFact.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import CatFact from "../components/CatFact/CatFact";
+import CatFact from "../../components/CatFact/CatFact";
 
 const CatFactPage = () => {
   return <CatFact />;
