+++ src/pages/DogImagePage/DogImage.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import DogImage from "../components/DogImage/DogImage";
+import DogImage from "../../components/DogImage/DogImage";
 
 const DogImagePage = () => {
   return <DogImage />;
