--- src/pages/HomePage/Homepage.css
@@ -1,16 +1,0 @@
-section {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 1em 0em;
-}
-
-section.shoppingListSection li {
-  list-style: none;
-}
-
-section.shoppingListSection ul {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
