+++ src/pages/ShoppingListPage/ShoppingList.css
@@ -1,0 +1,12 @@
+.shoppingListSection {
+  border: 2px solid black;
+  width: 50%;
+  margin: 1em auto;
+  grid-column-start: span 4;
+  background: rgba(255, 255, 255, 0.2);
+  border-radius: 16px;
+  box-shadow: 0 4px 30px rgba(0, 0, 0, 0.1);
+  backdrop-filter: blur(5px);
+  -webkit-backdrop-filter: blur(5px);
+  border: 1px solid rgba(255, 255, 255, 0.3);
+}
