--- src/pages/ShoppingListPage/ShoppingListPage.css
@@ -1,5 +1,0 @@
-.shoppingListSection {
-  border: 2px solid black;
-  width: 50%;
-  margin: 1em auto;
-}
